--- src/app/_components/Checkbox.stories.ts
@@ -1,17 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-import Checkbox from './Checkbox';
-
-const meta: Meta<typeof Checkbox> = {
-  title: 'Components/Checkbox',
-  component: Checkbox,
-  tags: ['autodocs'],
-} satisfies Meta<typeof Checkbox>;
-
-export default meta;
-type Story = StoryObj<typeof meta>;
-
-export const Example: Story = {
-  args: {
-    label: 'Label',
-  },
-};
